% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is_truthy}
\alias{is_truthy}
\title{Is an object TRUTHY?}
\usage{
is_truthy(object)
}
\arguments{
\item{object}{The object.}
}
\description{
Is an object TRUTHY?
}
\examples{
## Non-zero numbers are truthy
is_truthy(1)

## So are non-empty vectors that are not the zero scalar
is_truthy(c(1))

## Matrices are truthy, even if they are empty,
## because they are a class
is_truthy(matrix(nrow=0, ncol=0))

## But empty vectors are falsy
is_truthy(numeric())
}

