% Generated by roxygen2 (4.0.1): do not edit by hand
\name{not}
\alias{not}
\title{Negate the falsyness of a value.}
\usage{
not(rhs)
}
\arguments{
\item{rhs}{Value. If it is truthy, a falsy value (\code{FALSY}) is
returned, otherwise a truthy value (\code{TRUTHY}) is returned.}
}
\value{
Logical scalar.
}
\description{
Negate the falsyness of a value.
}
\examples{
not(FALSY)
not(TRUTHY)

## Check if directory is empty
tmp <- tempdir()
not(dir(tmp, all.files = TRUE, no.. = TRUE)) \%||\% message("Not empty")

cat("Hello!", file = tempfile())
not(dir(tmp, all.files = TRUE, no.. = TRUE)) \%||\% message("Not empty")
}

