% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsy-package.r
\docType{data}
\name{TRUTHY}
\alias{TRUTHY}
\title{A constant that is truthy.}
\format{An object of class \code{logical} of length 1.}
\usage{
TRUTHY
}
\description{
A constant that is truthy.
}
\examples{
is_truthy(TRUTHY)
is_falsy(FALSY)
}
\keyword{datasets}

