% Generated by roxygen2 (4.0.1): do not edit by hand
\name{try_quietly}
\alias{try_quietly}
\title{Try quietly}
\usage{
try_quietly(expr)
}
\arguments{
\item{expr}{The R expression to evaluate.}
}
\description{
We just call \code{try} and set its \code{silent} argument to
\code{TRUE}.
}
\examples{
\donttest{
try_quietly(library(foobar)) \%||\% message("Skipping foobar")
}
}

