% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Falsy package}
\alias{Falsy package}
\alias{Falsy package-package}
\title{Falsy package.}
\description{
We define a set of objects as \sQuote{falsy}. Every other values are
\sQuote{truthy}. This is inspired by other scripting languages like
Python and Ruby. See the \code{\%||\%} and the \code{\%&&\%} operators
on how to write concise conditional code using truthy and falsy values.
}
\details{
See \code{\link{is_falsy}} for the definition of what is \emph{falsy}
and what is \emph{truthy}.
}

