% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/falsy-package.r
\docType{data}
\name{FALSY}
\alias{FALSY}
\title{A constant that is falsy}
\format{\preformatted{ logi FALSE
}}
\usage{
FALSY
}
\description{
A constant that is falsy
}
\examples{
is_falsy(FALSY)
is_truthy(FALSY)
}
\keyword{datasets}

