% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/falsy-package.r
\name{or}
\alias{\%||\%}
\alias{or}
\title{Robust 'or' operator working with truthy and falsy values.}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left hand side, always evaluated.}

\item{rhs}{Right hand side, only evaluated if \code{lhs}
evaluates to FALSY.}
}
\description{
Robust 'or' operator working with truthy and falsy values.
}
\examples{
\donttest{
## Message if list is empty
l <- list(1,2,3)
l \%||\% message("list is empty")

l <- list()
l \%||\% message("list is empty")
}
}

