% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/falsy-package.r
\docType{data}
\name{TRUTHY}
\alias{TRUTHY}
\title{A constant that is truthy.}
\format{\preformatted{ logi TRUE
}}
\usage{
TRUTHY
}
\description{
A constant that is truthy.
}
\examples{
is_truthy(TRUTHY)
is_falsy(FALSY)
}
\keyword{datasets}

