% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/falsy-package.r
\name{try_quietly}
\alias{try_quietly}
\title{Try quietly}
\usage{
try_quietly(expr)
}
\arguments{
\item{expr}{The R expression to evaluate.}
}
\description{
We just call \code{try} and set its \code{silent} argument to
\code{TRUE}.
}

